#2709 refactored image with button component
